src/approov-protected-server/token-binding-check/requirements.txt
asgiref==3.2.10
Django==3.1.1
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2020.1
sqlparse==0.3.1
